GroupNo_6/backend/fastapi-ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
faiss-cpu==1.7.4
requests==2.31.0
openai==1.3.7
Pillow==10.1.0
imagehash==4.3.1
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
